좌우 조절 값 실시간 반영으로 수정

+++ include/device_app.h
@@ -20,8 +20,8 @@ typedef struct{
 }__attribute__((packed)) recvPacket;
 
 typedef struct{
-    int8_t fb;
-    int8_t lr;
+    int fb;
+    int lr;
 } mtVal;
 
 typedef struct{
